pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "meqpy"
description = "Python wrapper for MEQ."
version = "0.1.0"
license = {file = "LICENSE"}
authors = [{name = "MEQPY authors", email = "tamaranorman@google.com"}]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",

]
requires-python = ">=3.11"
dependencies = [
    "numpy >=1.26",
    "absl-py >= 2.0",
    "jax >=0.4",
    "oct2py >= 5.8.0"
]
readme = "README.md"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]